backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blis==1.3.0
CacheControl==0.14.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
elevenlabs==1.56.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.12
firebase==4.0.1
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-speech==2.31.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.10.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
groq==0.22.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
loguru==0.7.3
lxml==5.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.4
openai==1.73.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdfium2==4.30.1
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-http-client==3.3.7
python-multipart==0.0.20
python-vlc==3.0.21203
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
sendgrid==6.12.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starkbank-ecdsa==2.2.0
starlette==0.46.2
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.1
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yml==0.0.1
# Database dependencies
asyncpg==0.29.0
aiosqlite==0.20.0
sqlalchemy==2.0.36
alembic==1.14.0
# Configuration management
pyyaml==6.0.2
pydantic-settings==2.9.1
# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-timeout==2.4.2
pytest-mock==3.15.0
pytest-xdist==3.6.2
coverage==7.6.9
